refactor: test code에 상수 적용
test코드에서도 상수를 사용하도록 변경, Constant 폴더로 위치 변경
