+++ src/main/java/com/posomo/saltit/global/constant/ErrorMessage.java
@@ -1,4 +1,4 @@
-package com.posomo.saltit.global;
+package com.posomo.saltit.global.constant;
 
 public class ErrorMessage {
 	public static final String UNKNOWN_ERROR = "알 수 없는 에러가 발생했습니다.";
